+++ src/main/java/ui/controller/DeleteWorkOrder.java
@@ -12,15 +12,18 @@ public class DeleteWorkOrder extends RequestHandler {
     @Override
     public String handleRequest(HttpServletRequest request, HttpServletResponse response) {
         String workOrderId = request.getParameter("workOrderId");
-        int id = Integer.parseInt(workOrderId);
-        service.deleteWorkOrder(id);
-        HttpSession session = request.getSession();
-        Team team = (Team) session.getAttribute("userTeam");
-        request.setAttribute("teamLoggedIn", team);
-        Role role = (Role) session.getAttribute("userRole");
-        request.setAttribute("roleLoggedIn", role);
-        User user = (User) session.getAttribute("user");
-        request.setAttribute("userLoggedIn", user);
-        return "index.jsp";
+        if (workOrderId != null) {
+            int id = Integer.parseInt(workOrderId);
+            service.deleteWorkOrder(id);
+            HttpSession session = request.getSession();
+            Team team = (Team) session.getAttribute("userTeam");
+            request.setAttribute("teamLoggedIn", team);
+            Role role = (Role) session.getAttribute("userRole");
+            request.setAttribute("roleLoggedIn", role);
+            User user = (User) session.getAttribute("user");
+            request.setAttribute("userLoggedIn", user);
+            return "index.jsp";
+        }
+        return "deleteWorkOrder.jsp";
     }
 }
